requirements.txt
langchain
python-dotenv
pydantic
fastapi
uvicorn
jinja2
requests
google-generativeai

